+++ src/sound.h
@@ -27,7 +27,7 @@ typedef struct {
         often) in lq mode than in high-quality mode.  Maybe that
        should be fixed. :)
      */
-     void (*NeoFill)(int32 *Wave, int Count);
+     void (*NeoFill)(int16 *Wave, int Count);
      void (*HiFill)(void);
      void (*HiSync)(int32 ts);
 
@@ -41,10 +41,10 @@ extern int32 nesincsize;
 
 void SetSoundVariables(void);
 
-int GetSoundBuffer(int32 **W);
+int GetSoundBuffer(int16 **W);
 int FlushEmulateSound(void);
 extern int32 Wave[2048+512];
-extern int32 WaveFinal[2048+512];
+extern int16 WaveFinal[2048+512];
 extern int32 WaveHi[];
 extern uint32 soundtsinc;
